Spider/requirements.txt
PyExecJS==1.5.1
requests==2.18.4
Pillow==5.3.0
